♻️ K1E9GB prompt modules: decompose validation guards
