+++ .agentplane/tasks/202605290229-K1E9GB/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290229-K1E9GB`
+Title: Prompt module validation decomposition
+Canonical task record: `.agentplane/tasks/202605290229-K1E9GB/README.md`
+
+## Summary
+
+Prompt module validation decomposition
+
+Decompose packages/agentplane/src/runtime/prompt-modules/validation.ts by extracting prompt module validation constants and primitive field guards while preserving runtime prompt module, mutation set, and compiled graph validation behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/runtime/prompt-modules/validation.ts by extracting prompt module validation constants and primitive field guards while preserving runtime prompt module, mutation set, and compiled graph validation behavior.
+- Out of scope: unrelated refactors not required for "Prompt module validation decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T02:29:48.627Z
+- Branch: task/202605290229-K1E9GB/prompt-validation-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
