+++ .agentplane/tasks/202605290229-K1E9GB/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 K1E9GB task: Prompt module validation decomposition [202605290229-K1E9GB]
